Fix for screen rotation issue
When rotating screen while overlay vc is presented the underlying presenting vc gets corrupted after dismissal. I believe this is a bug in iOS11 and higher, because it does not happen on iOS10. Due to this, resetting the frame to screen size is more a workaround then a proper fix, feel free to suggest a better solution.